+++ www/SQLitePlugin.js
@@ -639,8 +639,10 @@
         name: SelfTest.DBNAME,
         location: 'default'
       }, function(db) {
-        return db.sqlBatch(['CREATE TABLE TestTable(TestColumn);', ['INSERT INTO TestTable (TestColumn) VALUES (?);', ['test-value']]], function() {
-          return db.executeSql('SELECT * FROM TestTable', [], function(resutSet) {
+        return db.sqlBatch(['CREATE TABLE TestTable(id integer primary key autoincrement unique, data);', ['INSERT INTO TestTable (data) VALUES (?);', ['test-value']]], function() {
+          var firstid;
+          firstid = -1;
+          return db.executeSql('SELECT id, data FROM TestTable', [], function(resutSet) {
             if (!resutSet.rows) {
               SelfTest.finishWithError(errorcb, 'Missing resutSet.rows');
               return;
@@ -653,49 +655,109 @@
               SelfTest.finishWithError(errorcb, "Incorrect resutSet.rows.length value: " + resutSet.rows.length + " (expected: 1)");
               return;
             }
-            if (!resutSet.rows.item(0).TestColumn) {
-              SelfTest.finishWithError(errorcb, 'Missing resutSet.rows.item(0).TestColumn');
+            if (resutSet.rows.item(0).id === void 0) {
+              SelfTest.finishWithError(errorcb, 'Missing resutSet.rows.item(0).id');
               return;
             }
-            if (resutSet.rows.item(0).TestColumn !== 'test-value') {
-              SelfTest.finishWithError(errorcb, "Incorrect resutSet.rows.item(0).TestColumn value: " + (resutSet.rows.item(0).TestColumn) + " (expected: 'test-value')");
+            firstid = resutSet.rows.item(0).id;
+            if (!resutSet.rows.item(0).data) {
+              SelfTest.finishWithError(errorcb, 'Missing resutSet.rows.item(0).data');
+              return;
+            }
+            if (resutSet.rows.item(0).data !== 'test-value') {
+              SelfTest.finishWithError(errorcb, "Incorrect resutSet.rows.item(0).data value: " + (resutSet.rows.item(0).data) + " (expected: 'test-value')");
               return;
             }
             return db.transaction(function(tx) {
-              return tx.executeSql('UPDATE TestTable SET TestColumn = ?', ['new-value']);
+              return tx.executeSql('UPDATE TestTable SET data = ?', ['new-value']);
             }, function(tx_err) {
               return SelfTest.finishWithError(errorcb, "UPDATE transaction error: " + tx_err);
             }, function() {
+              var readTransactionFinished;
+              readTransactionFinished = false;
               return db.readTransaction(function(tx2) {
-                return tx2.executeSql('SELECT * FROM TestTable', [], function(ignored, resutSet2) {
+                return tx2.executeSql('SELECT id, data FROM TestTable', [], function(ignored, resutSet2) {
                   if (!resutSet2.rows) {
-                    throw newSQLError('Missing resutSet.rows');
+                    throw newSQLError('Missing resutSet2.rows');
                   }
                   if (!resutSet2.rows.length) {
-                    throw newSQLError('Missing resutSet.rows.length');
+                    throw newSQLError('Missing resutSet2.rows.length');
                   }
                   if (resutSet2.rows.length !== 1) {
-                    throw newSQLError("Incorrect resutSet.rows.length value: " + resutSet.rows.length + " (expected: 1)");
+                    throw newSQLError("Incorrect resutSet2.rows.length value: " + resutSet2.rows.length + " (expected: 1)");
                   }
-                  if (!resutSet2.rows.item(0).TestColumn) {
-                    throw newSQLError('Missing resutSet.rows.item(0).TestColumn');
+                  if (!resutSet2.rows.item(0).id) {
+                    throw newSQLError('Missing resutSet2.rows.item(0).id');
                   }
-                  if (resutSet2.rows.item(0).TestColumn !== 'new-value') {
-                    throw newSQLError("Incorrect resutSet.rows.item(0).TestColumn value: " + (resutSet.rows.item(0).TestColumn) + " (expected: 'test-value')");
+                  if (resutSet2.rows.item(0).id !== firstid) {
+                    throw newSQLError("resutSet2.rows.item(0).id value " + (resutSet2.rows.item(0).id) + " does not match previous primary key id value (" + firstid + ")");
                   }
+                  if (!resutSet2.rows.item(0).data) {
+                    throw newSQLError('Missing resutSet2.rows.item(0).data');
+                  }
+                  if (resutSet2.rows.item(0).data !== 'new-value') {
+                    throw newSQLError("Incorrect resutSet2.rows.item(0).data value: " + (resutSet2.rows.item(0).data) + " (expected: 'test-value')");
+                  }
+                  return readTransactionFinished = true;
                 });
               }, function(tx2_err) {
                 return SelfTest.finishWithError(errorcb, "readTransaction error: " + tx2_err);
               }, function() {
-                return db.close(function() {
-                  return SQLiteFactory.deleteDatabase({
-                    name: SelfTest.DBNAME,
-                    location: 'default'
-                  }, successcb, function(cleanup_err) {
-                    return SelfTest.finishWithError(errorcb, "Cleanup error: " + cleanup_err);
+                if (!readTransactionFinished) {
+                  SelfTest.finishWithError(errorcb, 'readTransaction did not finish');
+                  return;
+                }
+                return db.transaction(function(tx3) {
+                  tx3.executeSql('DELETE FROM TestTable');
+                  return tx3.executeSql('INSERT INTO TestTable (data) VALUES(?)', [123]);
+                }, function(tx3_err) {
+                  return SelfTest.finishWithError(errorcb, "DELETE transaction error: " + tx3_err);
+                }, function() {
+                  var secondReadTransactionFinished;
+                  secondReadTransactionFinished = false;
+                  return db.readTransaction(function(tx4) {
+                    return tx4.executeSql('SELECT id, data FROM TestTable', [], function(ignored, resutSet3) {
+                      if (!resutSet3.rows) {
+                        throw newSQLError('Missing resutSet3.rows');
+                      }
+                      if (!resutSet3.rows.length) {
+                        throw newSQLError('Missing resutSet3.rows.length');
+                      }
+                      if (resutSet3.rows.length !== 1) {
+                        throw newSQLError("Incorrect resutSet3.rows.length value: " + resutSet3.rows.length + " (expected: 1)");
+                      }
+                      if (!resutSet3.rows.item(0).id) {
+                        throw newSQLError('Missing resutSet3.rows.item(0).id');
+                      }
+                      if (resutSet3.rows.item(0).id === firstid) {
+                        throw newSQLError("resutSet3.rows.item(0).id value " + (resutSet3.rows.item(0).id) + " incorrectly matches previous unique key id value value (" + firstid + ")");
+                      }
+                      if (!resutSet3.rows.item(0).data) {
+                        throw newSQLError('Missing resutSet3.rows.item(0).data');
+                      }
+                      if (resutSet3.rows.item(0).data !== 123) {
+                        throw newSQLError("Incorrect resutSet3.rows.item(0).data value: " + (resutSet3.rows.item(0).data) + " (expected 123)");
+                      }
+                      return secondReadTransactionFinished = true;
+                    });
+                  }, function(tx4_err) {
+                    return SelfTest.finishWithError(errorcb, "second readTransaction error: " + tx4_err);
+                  }, function() {
+                    if (!secondReadTransactionFinished) {
+                      SelfTest.finishWithError(errorcb, 'second readTransaction did not finish');
+                      return;
+                    }
+                    return db.close(function() {
+                      return SQLiteFactory.deleteDatabase({
+                        name: SelfTest.DBNAME,
+                        location: 'default'
+                      }, successcb, function(cleanup_err) {
+                        return SelfTest.finishWithError(errorcb, "Cleanup error: " + cleanup_err);
+                      });
+                    }, function(close_err) {
+                      return SelfTest.finishWithError(errorcb, "close error: " + close_err);
+                    });
                   });
-                }, function(close_err) {
-                  return SelfTest.finishWithError(errorcb, "close error: " + close_err);
                 });
               });
             });